+++ src/memory/vector.cpp
@@ -4,19 +4,19 @@
 #include "include/vector.hh"
 
 namespace VFEngine {
-    Vector::Vector() : _vector(std::vector<int32_t>(State::MAX_VECTOR_SIZE, 0)),
+    Vector::Vector() : _vector(std::vector<uint64_t>(State::MAX_VECTOR_SIZE, 0)),
                        _state(std::make_shared<State>(State::MAX_VECTOR_SIZE)) {
     }
 
-    Vector::Vector(const std::vector<int32_t> &vector) : _vector(vector),
+    Vector::Vector(const std::vector<uint64_t> &vector) : _vector(vector),
                                                          _state(std::make_shared<State>(
                                                              std::min(vector.size(), State::MAX_VECTOR_SIZE))) {
     }
 
-    Vector::Vector(const int32_t &size) : _vector(std::vector<int32_t>(std::min((int32_t) State::MAX_VECTOR_SIZE, size),
+    Vector::Vector(const int32_t &size) : _vector(std::vector<uint64_t>(std::min(static_cast<int32_t>(State::MAX_VECTOR_SIZE), size),
                                                                        0)),
                                           _state(std::make_shared<State>(
-                                              std::min(size, (int32_t) State::MAX_VECTOR_SIZE))) {
+                                              std::min(size, static_cast<int32_t>(State::MAX_VECTOR_SIZE)))) {
     }
